@HCLarsen 2021-12-20
After merging, I've found that I've had to revert it to the previous state, as it's failing the test `test_should_return_the_specified_number_of_passwords`.

Please check that the entire test suite is passing before generating a PR.